Drop the library search hack for running the tests
As this is a rather old feedstock, it relied on a hack on macOS and
Linux to ensure libraries were loaded correctly when running the tests
as `conda-build` had not worked it's prefix magic to fix the paths in
binaries to load them from the correct location. The workaround was to
set the library fallback paths for loading these libraries to search in
the correct location when nothing else worked. However SIP on macOS
10.11 made this no longer a viable solution. As the `compiler` syntax
pulls in a fix that sets the `rpath` correctly on macOS and Linux when
building libraries, there is no need for this hack to search for the
linked dependencies as they will already be found correctly using the
`rpath`. So just drop the hack. Of course `conda-build` will still do
binary prefix substitution, which is fine as well.