Drop explicit pinnings of dependencies
With `conda-build` 3, `conda-smithy` 3, and `conda-forge-pinning`, these
version constraints of dependencies are pulled in by re-rendering from a
global pinning file and then applied to the recipe. Thus there is no
need for every recipe to carry around this information explicitly. Hence
go ahead and drop these pinnings to simplify the recipe and improve
maintenance going forward.