Apply upstream patch to fix macOS UUID issue
On macOS, the system supplies UUID, which we would prefer to use. It
also doesn't have `.pc` (pkg-config files) to specify where to look,
which complicates matters. Instead of using the hacky workaround we had
before to deal with this issue, apply an upstream patch to fix the issue
and specify that UUID should not be looked for elsewhere.